+++ R/link_eqtl.R
@@ -11,9 +11,10 @@
 #' allowed, default = "CEU". Passed to `LDlinkR::LDexpress()`.
 #' @param r2d Either "r2" for LD r^2 or "d" for LD D', default = "r2". Passed 
 #' to `LDlinkR::LDexpress()`.
-#' @param LDtoken Personal access token for accessing 1000 Genomes LD data via 
+#' @param token Personal access token for accessing 1000 Genomes LD data via 
 #' LDlink API. See `LDlinkR` package documentation.
-#' @param ... Optional arguments passed to `LDlinkR::LDexpress()`
+#' @param ... Optional arguments such as `genome_build` which are passed on to
+#'   `LDlinkR::LDexpress()`
 #' @return Returns an object of class 'locus' with an extra list element 'LDexp'
 #'   containing a dataframe of information obtained via `LDexpress()`.
 #' @details
@@ -28,7 +29,7 @@
 link_eqtl <- function(loc,
                       pop = "CEU",
                       r2d = "r2",
-                      LDtoken = "", ...) {
+                      token = "", ...) {
   if (!inherits(loc, "locus")) stop("Not a locus object")
   if (!requireNamespace("LDlinkR", quietly = TRUE)) {
     stop("Package 'LDlinkR' must be installed to use this feature",
@@ -37,9 +38,9 @@ link_eqtl <- function(loc,
   labs <- loc$labs
   index_snp <- loc$index_snp
   
-  if (LDtoken == "") stop("LDtoken is missing")
+  if (token == "") stop("token is missing")
   LDexp <- mem_LDexpress(snps = index_snp, pop = pop, r2d = r2d, 
-                         token = LDtoken, ...)
+                         token = token, ...)
   for (i in c("R2", "D'", "Effect_Size", "P_value")) {
     LDexp[, i] <- as.numeric(LDexp[, i])
   }